Optimizations

+++ StoryboardSystem/Utility/SerializationUtility.cs
@@ -16,23 +16,23 @@ internal static class SerializationUtility {
     public static bool TrySerialize(object obj, BinaryWriter writer) {
         switch (obj) {
             case bool val:
-                writer.Write((int) SerializableType.Bool);
+                writer.Write((byte) SerializableType.Bool);
                 writer.Write(val);
                 return true;
             case int val:
-                writer.Write((int) SerializableType.Int);
+                writer.Write((byte) SerializableType.Int);
                 writer.Write(val);
                 return true;
             case float val:
-                writer.Write((int) SerializableType.Float);
+                writer.Write((byte) SerializableType.Float);
                 writer.Write(val);
                 return true;
             case string val:
-                writer.Write((int) SerializableType.String);
+                writer.Write((byte) SerializableType.String);
                 writer.Write(val);
                 return true;
             case object[] arr:
-                writer.Write((int) SerializableType.Array);
+                writer.Write((byte) SerializableType.Array);
                 writer.Write(arr.Length);
 
                 foreach (object val in arr) {
@@ -47,7 +47,7 @@ internal static class SerializationUtility {
     }
 
     public static bool TryDeserialize(BinaryReader reader, out object obj) {
-        switch ((SerializableType) reader.ReadInt32()) {
+        switch ((SerializableType) reader.ReadByte()) {
             case SerializableType.Null:
                 obj = null;
                 return true;
@@ -70,6 +70,7 @@ internal static class SerializationUtility {
                 for (int i = 0; i < length; i++) {
                     if (TryDeserialize(reader, out arr[i]))
                         continue;
+                    
                     obj = null;
 
                     return false;
